tank_backend/requirements.txt
Django>=3.0.6
djangorestframework>=3.11.0
django-cors-headers>=3.3.0
psycopg2>=2.8.6